Separate Zoom helper funcs from tool
