+++ toolkits/zoom/arcade_zoom/tools/constants.py
@@ -1,0 +1,1 @@
+ZOOM_BASE_URL = "https://api.zoom.us/v2"
